requirements.txt
tabulate==0.9.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10